# api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Auth API API",
        "title": "Auth API",
        "contact": {
            "name": "Nathan Faucett",
            "email": "nathanfaucett@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        }
    },
    "basePath": "/",
    "paths": {
...
        "/applications": {
            "get": {
                "security": [
                    {
                        "Authorization": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "application"
                ],
                "summary": "Get applications",
                "operationId": "applications",
                "parameters": [
                    {
                        "type": "integer",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Pagination-Application"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/Errors"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/Errors"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/Errors"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/Errors"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "Authorization": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "application"
                ],
                "summary": "Create application",
                "operationId": "create-application",
                "parameters": [
                    {
                        "description": "create application",
                        "name": "application",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/CreateApplication"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/Application"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/Errors"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/Errors"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/Errors"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/Errors"
                        }
                    }
                }
            }
        },
        "/applications/{applicationId}/tenents": {
            "get": {
                "security": [
                    {
                        "Authorization": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tenent"
                ],
                "summary": "Get application tenents",
                "operationId": "tenents",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "application id",
                        "name": "applicationId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Pagination-Tenent"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/Errors"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/Errors"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/Errors"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/Errors"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "Authorization": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tenent"
                ],
                "summary": "Create application tenent",
                "operationId": "create-tenent",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "application id",
                        "name": "applicationId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "create application",
                        "name": "tenent",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/CreateTenent"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/Tenent"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/Errors"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/Errors"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/Errors"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/Errors"
                        }
                    }
                }
            }
        },
        "/applications/{applicationId}/tenents/{id}": {
            "get": {
                "security": [
                    {
                        "Authorization": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tenent"
                ],
                "summary": "Get application tenent by id",
                "operationId": "tenent",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "application id",
                        "name": "applicationId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "application tenent id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Tenent"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/Errors"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/Errors"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/Errors"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/Errors"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "Authorization": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tenent"
                ],
                "summary": "Delete application tenent",
                "operationId": "delete-tenent",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "application id",
                        "name": "applicationId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "application tenent id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/Errors"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/Errors"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/Errors"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/Errors"
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "Authorization": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tenent"
                ],
                "summary": "Update application tenent",
                "operationId": "update-tenent",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "application id",
                        "name": "applicationId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "application tenent id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "update application",
                        "name": "tenent",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/UpdateTenent"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Tenent"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/Errors"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/Errors"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/Errors"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/Errors"
                        }
                    }
                }
            }
        },
        "/applications/{applicationId}/tenents/{id}/private-key": {
            "get": {
                "security": [
                    {
                        "Authorization": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tenent"
                ],
                "summary": "Get application tenent by id",
                "operationId": "tenent-private-key",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "application id",
                        "name": "applicationId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "application tenent id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/Errors"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/Errors"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/Errors"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/Errors"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "Application": {
            "type": "object",
            "required": [
                "created_at",
                "description",
                "id",
                "updated_at",
                "uri"
            ],
            "properties": {
                "created_at": {
                    "type": "string",
                    "format": "date-time"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string",
                    "format": "date-time"
                },
                "uri": {
                    "type": "string"
                },
                "website": {
                    "type": "string"
                }
            }
        },
...
        "CreateApplication": {
            "type": "object",
            "required": [
                "description",
                "uri"
            ],
            "properties": {
                "description": {
                    "type": "string"
                },
                "uri": {
                    "type": "string"
                }
            }
        },
...
        "CreateTenent": {
            "type": "object",
            "required": [
                "description",
                "uri"
            ],
            "properties": {
                "algorithm": {
                    "type": "string"
                },
                "authorization_website": {
                    "type": "string"
                },
                "client_id": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "email_endpoint": {
                    "type": "string"
                },
                "expires_in_seconds": {
                    "type": "integer"
                },
                "password_reset_expires_in_seconds": {
                    "type": "integer"
                },
                "phone_number_endpoint": {
                    "type": "string"
                },
                "private_key": {
                    "type": "string"
                },
                "public_key": {
                    "type": "string"
                },
                "refresh_expires_in_seconds": {
                    "type": "integer"
                },
                "registration_website": {
                    "type": "string"
                },
                "uri": {
                    "type": "string"
                }
            }
        },
...
        "Errors": {
            "type": "object",
            "required": [
                "errors"
            ],
            "properties": {
                "errors": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "array",
                        "items": {
                            "$ref": "#/definitions/ErrorMessage"
                        }
                    }
                }
            }
        },
...
        "Pagination-Application": {
            "type": "object",
            "required": [
                "has_more",
                "items"
            ],
            "properties": {
                "has_more": {
                    "type": "boolean"
                },
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/Application"
                    }
                }
            }
        },
        "Pagination-Tenent": {
            "type": "object",
            "required": [
                "has_more",
                "items"
            ],
            "properties": {
                "has_more": {
                    "type": "boolean"
                },
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/Tenent"
                    }
                }
            }
        },
...
        "Tenent": {
            "type": "object",
            "required": [
                "algorithm",
                "application_id",
                "authorization_website",
                "client_id",
                "created_at",
                "description",
                "expires_in_seconds",
                "id",
                "password_reset_expires_in_seconds",
                "refresh_expires_in_seconds",
                "updated_at",
                "uri"
            ],
            "properties": {
                "algorithm": {
                    "type": "string"
                },
                "application_id": {
                    "type": "integer"
                },
                "authorization_website": {
                    "type": "string"
                },
                "client_id": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string",
                    "format": "date-time"
                },
                "description": {
                    "type": "string"
                },
                "expires_in_seconds": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "password_reset_expires_in_seconds": {
                    "type": "integer"
                },
                "public_key": {
                    "type": "string"
                },
                "refresh_expires_in_seconds": {
                    "type": "integer"
                },
                "registration_website": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string",
                    "format": "date-time"
                },
                "uri": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "Authorization": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "Locale": {
            "type": "apiKey",
            "name": "X-Locale",
            "in": "header"
        },
        "TenentId": {
            "type": "apiKey",
            "name": "Tenent-Id",
            "in": "header"
        },
        "Timezone": {
            "type": "apiKey",
            "name": "X-Timezone",
            "in": "header"
        }
    }
}